Changed event description and added closing
I have added the proper event descriptions that users need when hovering
over an event, also added that if the event is duration then both start
date and end date should be shown and if it is non-duration then only
start date should be shown. Another new change is that i have added the
manual closing and removed the closing on when not hovered over event.
This allows the users to use the edit and delete button that i have
added this iteration to do the respective tasks.